Split up fields into different field types for Painless Fields API
