+++ server/src/main/java/org/elasticsearch/script/field/DateMillisField.java
@@ -10,7 +10,45 @@ package org.elasticsearch.script.field;
 
 import org.elasticsearch.script.JodaCompatibleZonedDateTime;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class DateMillisField extends Field<JodaCompatibleZonedDateTime> {
+
+    /* ---- Conversion Helpers To Other Fields ---- */
+
+    public static LongField convertDateMillisToLongField(DateMillisField sourceField) {
+        FieldValues<JodaCompatibleZonedDateTime> fv = sourceField.getFieldValues();
+        return new LongField(sourceField.getName(), new DelegatingFieldValues<Long, JodaCompatibleZonedDateTime>(fv) {
+            @Override
+            public List<Long> getValues() {
+                return values.getValues().stream().map(dt -> dt.toInstant().toEpochMilli()).collect(Collectors.toList());
+            }
+
+            @Override
+            public Long getNonPrimitiveValue() {
+                return convertDateMillisToLong(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public long getLongValue() {
+                return convertDateMillisToLong(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public double getDoubleValue() {
+                return convertDateMillisToLong(values.getNonPrimitiveValue());
+            }
+        });
+    }
+
+    /* ---- Conversion Helpers To Other Types ---- */
+
+    public static long convertDateMillisToLong(JodaCompatibleZonedDateTime dt) {
+        return dt.toInstant().toEpochMilli();
+    }
+
+    /* ---- DateMillis Field Members ---- */
 
     public DateMillisField(String name, FieldValues<JodaCompatibleZonedDateTime> values) {
         super(name, values);
