+++ server/src/main/java/org/elasticsearch/script/field/DateNanosField.java
@@ -10,7 +10,51 @@ package org.elasticsearch.script.field;
 
 import org.elasticsearch.script.JodaCompatibleZonedDateTime;
 
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class DateNanosField extends Field<JodaCompatibleZonedDateTime> {
+
+    /* ---- Conversion Helpers To Other Fields ---- */
+
+    public static LongField convertDateNanosToLongField(DateNanosField sourceField) {
+        FieldValues<JodaCompatibleZonedDateTime> fv = sourceField.getFieldValues();
+        return new LongField(sourceField.getName(), new DelegatingFieldValues<Long, JodaCompatibleZonedDateTime>(fv) {
+            protected long nanoLong(JodaCompatibleZonedDateTime dt) {
+                return ChronoUnit.NANOS.between(Instant.EPOCH, dt.toInstant());
+            }
+
+            @Override
+            public List<Long> getValues() {
+                return values.getValues().stream().map(this::nanoLong).collect(Collectors.toList());
+            }
+
+            @Override
+            public Long getNonPrimitiveValue() {
+                return convertDateNanosToLong(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public long getLongValue() {
+                return convertDateNanosToLong(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public double getDoubleValue() {
+                return convertDateNanosToLong(values.getNonPrimitiveValue());
+            }
+        });
+    }
+
+    /* ---- Conversion Helpers To Other Types ---- */
+
+    public static long convertDateNanosToLong(JodaCompatibleZonedDateTime dt) {
+        return ChronoUnit.NANOS.between(Instant.EPOCH, dt.toInstant());
+    }
+
+    /* ---- DateNanos Field Members ---- */
 
     public DateNanosField(String name, FieldValues<JodaCompatibleZonedDateTime> values) {
         super(name, values);
