+++ server/src/main/java/org/elasticsearch/script/field/DoubleField.java
@@ -8,7 +8,52 @@
 
 package org.elasticsearch.script.field;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class DoubleField extends Field<Double> {
+
+    /* ---- Conversion Helpers To Other Fields ---- */
+
+    public static BigIntegerField convertDoubleToBigIntegerField(DoubleField sourceField) {
+        FieldValues<Double> fv = sourceField.getFieldValues();
+        return new BigIntegerField(sourceField.getName(), new DelegatingFieldValues<java.math.BigInteger, Double>(fv) {
+            @Override
+            public List<BigInteger> getValues() {
+                return values.getValues().stream().map(DoubleField::convertDoubleToBigInteger).collect(Collectors.toList());
+            }
+
+            @Override
+            public BigInteger getNonPrimitiveValue() {
+                return convertDoubleToBigInteger(values.getDoubleValue());
+            }
+        });
+    }
+
+    public static LongField convertDoubleToLongField(DoubleField sourceField) {
+        FieldValues<Double> fv = sourceField.getFieldValues();
+        return new LongField(sourceField.getName(), new DelegatingFieldValues<Long, Double>(fv) {
+            @Override
+            public List<Long> getValues() {
+                return values.getValues().stream().map(Double::longValue).collect(Collectors.toList());
+            }
+
+            @Override
+            public Long getNonPrimitiveValue() {
+                return values.getLongValue();
+            }
+        });
+    }
+
+    /* ---- Conversion Helpers To Other Types ---- */
+
+    public static BigInteger convertDoubleToBigInteger(double dbl) {
+        return BigDecimal.valueOf(dbl).toBigInteger();
+    }
+
+    /* ---- Double Field Members ---- */
 
     public DoubleField(String name, FieldValues<Double> values) {
         super(name, values);
