+++ server/src/main/java/org/elasticsearch/script/field/EmptyField.java
@@ -15,6 +15,9 @@ import java.util.List;
  * Script field with no mapping, always returns {@code defaultValue}.
  */
 public class EmptyField<T> extends Field<T> {
+
+    /* ---- Empty Field Members ---- */
+
     public EmptyField(String name) {
         super(name, null);
     }
