+++ server/src/main/java/org/elasticsearch/script/field/LongField.java
@@ -8,7 +8,86 @@
 
 package org.elasticsearch.script.field;
 
+import java.math.BigInteger;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class LongField extends Field<Long> {
+
+    /* ---- Conversion Class From Other Fields ----*/
+
+    /**
+     * Convert to a {@link LongField} from Double, String, DateMillis, DateNanos, BigInteger or Boolean Fields.
+     * Double is cast to a Long.
+     * String is parsed as a Long.
+     * DateMillis is milliseconds since epoch.
+     * DateNanos is nanoseconds since epoch.
+     * {@link java.math.BigInteger#longValue()} is used for the BigInteger conversion.
+     * Boolean is {@code 1L} if {@code true}, {@code 0L} if {@code false}.
+     */
+    public static final Converter<Long, LongField> Long;
+
+    static {
+        Long = new Converter<Long, LongField>() {
+            @Override
+            public LongField convert(Field<?> sourceField) {
+                if (sourceField instanceof DoubleField) {
+                    return DoubleField.convertDoubleToLongField((DoubleField) sourceField);
+                }
+
+                if (sourceField instanceof StringField) {
+                    return StringField.convertStringToLongField((StringField) sourceField);
+                }
+
+                if (sourceField instanceof DateMillisField) {
+                    return DateMillisField.convertDateMillisToLongField((DateMillisField) sourceField);
+                }
+
+                if (sourceField instanceof DateNanosField) {
+                    return DateNanosField.convertDateNanosToLongField((DateNanosField) sourceField);
+                }
+
+                if (sourceField instanceof BigIntegerField) {
+                    return BigIntegerField.convertBigIntegerToLongField((BigIntegerField) sourceField);
+                }
+
+                if (sourceField instanceof BooleanField) {
+                    return BooleanField.convertBooleanToLongField((BooleanField) sourceField);
+                }
+
+                throw new InvalidConversion(sourceField.getClass(), getFieldClass());
+            }
+
+            @Override
+            public Class<LongField> getFieldClass() {
+                return LongField.class;
+            }
+
+            @Override
+            public Class<Long> getTargetClass() {
+                return Long.class;
+            }
+        };
+    }
+
+    /* ---- Conversion Helpers To Other Fields ---- */
+
+    public static BigIntegerField convertLongToBigIntegerField(LongField sourceField) {
+        FieldValues<Long> fv = sourceField.getFieldValues();
+        return new BigIntegerField(sourceField.getName(), new DelegatingFieldValues<BigInteger, Long>(fv) {
+            @Override
+            public List<BigInteger> getValues() {
+                return values.getValues().stream().map(BigInteger::valueOf).collect(Collectors.toList());
+            }
+
+            @Override
+            public BigInteger getNonPrimitiveValue() {
+                return BigInteger.valueOf(values.getLongValue());
+            }
+        });
+    }
+
+    /* ---- Long Field Members ---- */
 
     public LongField(String name, FieldValues<Long> values) {
         super(name, values);
