+++ server/src/main/java/org/elasticsearch/script/field/StringField.java
@@ -8,7 +8,86 @@
 
 package org.elasticsearch.script.field;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class StringField extends Field<String> {
+
+    /* ---- Conversion Helpers To Other Fields ---- */
+
+    public static BigIntegerField convertStringToBigIntegerField(StringField sourceField) {
+        FieldValues<String> fv = sourceField.getFieldValues();
+        return new BigIntegerField(sourceField.getName(), new DelegatingFieldValues<java.math.BigInteger, String>(fv) {
+            @Override
+            public List<BigInteger> getValues() {
+                // This may throw NumberFormatException, should we catch and truncate the List? (#76951)
+                return values.getValues().stream().map(StringField::convertStringToBigInteger).collect(Collectors.toList());
+            }
+
+            @Override
+            public BigInteger getNonPrimitiveValue() {
+                return convertStringToBigInteger(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public long getLongValue() {
+                return getNonPrimitiveValue().longValue();
+            }
+
+            @Override
+            public double getDoubleValue() {
+                return getNonPrimitiveValue().doubleValue();
+            }
+        });
+    }
+
+    public static LongField convertStringToLongField(StringField sourceField) {
+        FieldValues<String> fv = sourceField.getFieldValues();
+        return new LongField(sourceField.getName(), new DelegatingFieldValues<Long, String>(fv) {
+            @Override
+            public List<Long> getValues() {
+                return values.getValues().stream().map(StringField::convertStringToLong).collect(Collectors.toList());
+            }
+
+            @Override
+            public Long getNonPrimitiveValue() {
+                return convertStringToLong(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public long getLongValue() {
+                return convertStringToLong(values.getNonPrimitiveValue());
+            }
+
+            @Override
+            public double getDoubleValue() {
+                // conversion is to LongField, doesn't make sense to parse a Double out of the String here.
+                return convertStringToLong(values.getNonPrimitiveValue());
+            }
+        });
+    }
+
+    /* ---- Conversion Helpers To Other Types ---- */
+
+    public static BigInteger convertStringToBigInteger(String str) {
+        try {
+            return new BigInteger(str);
+        } catch (NumberFormatException e) {
+            return new BigDecimal(str).toBigInteger();
+        }
+    }
+
+    public static long convertStringToLong(String str) {
+        return Long.parseLong(str);
+    }
+
+    public static double convertStringToDouble(String str) {
+        return Double.parseDouble(str);
+    }
+
+    /* ---- String Field Members ---- */
 
     public StringField(String name, FieldValues<String> values) {
         super(name, values);
