+++ server/src/test/java/org/elasticsearch/script/field/ConvertersTests.java
@@ -56,7 +56,7 @@ public class ConvertersTests extends ESTestCase {
             }
         });
 
-        Field<BigInteger> dst = src.as(Field.BigInteger);
+        Field<BigInteger> dst = src.as(BigIntegerField.BigInteger);
 
         List<BigInteger> expected = LongStream.of(raw).mapToObj(BigInteger::valueOf).collect(Collectors.toList());
         assertEquals(expected, dst.getValues());
@@ -65,7 +65,7 @@ public class ConvertersTests extends ESTestCase {
         assertEquals(raw[0], dst.getLong(10));
         assertEquals((double) raw[0], dst.getDouble(10.0d), 0.1d);
 
-        Field<Long> dstLong = dst.as(Field.Long);
+        Field<Long> dstLong = dst.as(LongField.Long);
         assertEquals(LongStream.of(raw).boxed().collect(Collectors.toList()), dstLong.getValues());
         assertEquals(Long.valueOf(raw[0]), dstLong.getValue(null));
         assertEquals(raw[0], dstLong.getLong(10));
@@ -107,7 +107,7 @@ public class ConvertersTests extends ESTestCase {
             }
         });
 
-        Field<BigInteger> dst = src.as(Field.BigInteger);
+        Field<BigInteger> dst = src.as(BigIntegerField.BigInteger);
         BigInteger maxDouble = new BigInteger("17976931348623157" + "0".repeat(292));
         List<BigInteger> expected = List.of(maxDouble, BigInteger.ZERO, new BigInteger("34028234663852886" + "0".repeat(23)),
                                             BigInteger.ZERO, BigInteger.ZERO,
@@ -118,7 +118,7 @@ public class ConvertersTests extends ESTestCase {
         assertEquals(Long.MAX_VALUE, dst.getLong(10));
         assertEquals(Double.MAX_VALUE, dst.getDouble(10.0d), 0.1d);
 
-        Field<Long> lng = src.as(Field.Long);
+        Field<Long> lng = src.as(LongField.Long);
         List<Long> lngExpected = List.of(Long.MAX_VALUE, 0L, Long.MAX_VALUE, 0L, 0L, Long.MAX_VALUE, Long.MIN_VALUE);
         assertEquals(lngExpected, lng.getValues());
         assertEquals(Long.valueOf(Long.MAX_VALUE), lng.getValue(null));
@@ -131,7 +131,7 @@ public class ConvertersTests extends ESTestCase {
                                    Double.MIN_VALUE + "");
         Field<String> src = new StringField("", new ListFieldValues<>(raw));
 
-        Field<BigInteger> dst = src.as(Field.BigInteger);
+        Field<BigInteger> dst = src.as(BigIntegerField.BigInteger);
         BigInteger maxDouble = new BigInteger("17976931348623157" + "0".repeat(292));
         List<BigInteger> expected = List.of(new BigInteger(raw.get(0)), new BigInteger(raw.get(1)), new BigInteger(raw.get(2)), maxDouble,
                                             BigInteger.ZERO);
@@ -146,7 +146,7 @@ public class ConvertersTests extends ESTestCase {
         List<String> raw = List.of(rand + "", Long.MAX_VALUE + "", Long.MIN_VALUE + "", "0", "100");
         Field<String> src = new StringField("", new ListFieldValues<>(raw));
 
-        Field<Long> dst = src.as(Field.Long);
+        Field<Long> dst = src.as(LongField.Long);
         assertEquals(List.of(rand, Long.MAX_VALUE, Long.MIN_VALUE, 0L, 100L), dst.getValues());
         assertEquals(Long.valueOf(rand), dst.getValue(null));
         assertEquals(rand, dst.getLong(10)); // overflow
@@ -157,13 +157,13 @@ public class ConvertersTests extends ESTestCase {
         List<Boolean> raw = List.of(Boolean.TRUE, Boolean.FALSE);
         Field<Boolean> src = new BooleanField("", new ListFieldValues<>(raw));
 
-        Field<BigInteger> dst = src.as(Field.BigInteger);
+        Field<BigInteger> dst = src.as(BigIntegerField.BigInteger);
         assertEquals(List.of(BigInteger.ONE, BigInteger.ZERO), dst.getValues());
         assertEquals(BigInteger.ONE, dst.getValue(null));
         assertEquals(1L, dst.getLong(10L));
         assertEquals(1.0d, dst.getDouble(1234.0d), 0.1d);
 
-        Field<Long> dstLong = src.as(Field.Long);
+        Field<Long> dstLong = src.as(LongField.Long);
         assertEquals(List.of(1L, 0L), dstLong.getValues());
         assertEquals(Long.valueOf(1), dstLong.getValue(null));
         assertEquals(1L, dstLong.getLong(10L));
@@ -171,14 +171,14 @@ public class ConvertersTests extends ESTestCase {
 
         List<Boolean> rawRev = List.of(Boolean.FALSE, Boolean.TRUE);
         src = new BooleanField("", new ListFieldValues<>(rawRev));
-        dst = src.as(Field.BigInteger);
+        dst = src.as(BigIntegerField.BigInteger);
 
         assertEquals(List.of(BigInteger.ZERO, BigInteger.ONE), dst.getValues());
         assertEquals(BigInteger.ZERO, dst.getValue(null));
         assertEquals(0L, dst.getLong(10L));
         assertEquals(0.0d, dst.getDouble(1234.0d), 0.1d);
 
-        dstLong = src.as(Field.Long);
+        dstLong = src.as(LongField.Long);
         assertEquals(List.of(0L, 1L), dstLong.getValues());
         assertEquals(Long.valueOf(0), dstLong.getValue(null));
         assertEquals(0L, dstLong.getLong(10L));
@@ -187,10 +187,10 @@ public class ConvertersTests extends ESTestCase {
 
     public void testInvalidFieldConversion() {
         Field<GeoPoint> src = new GeoPointField("", new ListFieldValues<>(List.of(new GeoPoint(0, 0))));
-        InvalidConversion ic = expectThrows(InvalidConversion.class, () -> src.as(Field.BigInteger));
+        InvalidConversion ic = expectThrows(InvalidConversion.class, () -> src.as(BigIntegerField.BigInteger));
         assertEquals("Cannot convert from [GeoPointField] using converter [BigIntegerField]", ic.getMessage());
 
-        ic = expectThrows(InvalidConversion.class, () -> src.as(Field.Long));
+        ic = expectThrows(InvalidConversion.class, () -> src.as(LongField.Long));
         assertEquals("Cannot convert from [GeoPointField] using converter [LongField]", ic.getMessage());
     }
 
@@ -205,13 +205,13 @@ public class ConvertersTests extends ESTestCase {
         Field<JodaCompatibleZonedDateTime> src = new DateMillisField("", new ListFieldValues<>(raw));
 
         List<BigInteger> expectedBigInteger = LongStream.of(rawMilli).mapToObj(BigInteger::valueOf).collect(Collectors.toList());
-        Field<BigInteger> dstBigInteger = src.as(Field.BigInteger);
+        Field<BigInteger> dstBigInteger = src.as(BigIntegerField.BigInteger);
         assertEquals(expectedBigInteger, dstBigInteger.getValues());
         assertEquals(expectedBigInteger.get(0), dstBigInteger.getValue(null));
         assertEquals(rawMilli[0], dstBigInteger.getLong(-1000L));
         assertEquals((double) rawMilli[0], dstBigInteger.getDouble(-1234.5d), 1.1d);
 
-        Field<Long> dstLong = src.as(Field.Long);
+        Field<Long> dstLong = src.as(LongField.Long);
         assertEquals(LongStream.of(rawMilli).boxed().collect(Collectors.toList()), dstLong.getValues());
         assertEquals(LongStream.of(rawMilli).boxed().collect(Collectors.toList()), dstLong.getValues());
         assertEquals(Long.valueOf(rawMilli[0]), dstLong.getValue(-100L));
@@ -230,13 +230,13 @@ public class ConvertersTests extends ESTestCase {
         Field<JodaCompatibleZonedDateTime> src = new DateNanosField("", new ListFieldValues<>(raw));
 
         List<BigInteger> expectedBigInteger = LongStream.of(rawNanos).mapToObj(BigInteger::valueOf).collect(Collectors.toList());
-        Field<BigInteger> dstBigInteger = src.as(Field.BigInteger);
+        Field<BigInteger> dstBigInteger = src.as(BigIntegerField.BigInteger);
         assertEquals(expectedBigInteger, dstBigInteger.getValues());
         assertEquals(expectedBigInteger.get(0), dstBigInteger.getValue(null));
         assertEquals(rawNanos[0], dstBigInteger.getLong(-1000L));
         assertEquals((double) rawNanos[0], dstBigInteger.getDouble(-1234.5d), 1.1d);
 
-        Field<Long> dstLong = src.as(Field.Long);
+        Field<Long> dstLong = src.as(LongField.Long);
         assertEquals(LongStream.of(rawNanos).boxed().collect(Collectors.toList()), dstLong.getValues());
         assertEquals(LongStream.of(rawNanos).boxed().collect(Collectors.toList()), dstLong.getValues());
         assertEquals(Long.valueOf(rawNanos[0]), dstLong.getValue(-100L));
